requirements.txt
streamlit
python-dotenv
deepgram-sdk == 3.2.0
httpx
fastapi
uvicorn
pydantic
requests
google-generativeai
langchain
langchain_core
langchain_google_genai
whisperX
